+++ src/app/[locale]/dashboard/export/page.tsx
@@ -1,4 +1,4 @@
-import { ExportPagePanel } from "@/components/passwords/export-dialog";
+import { ExportPagePanel } from "@/components/passwords/password-export";
 
 export default function ExportPage() {
   return <ExportPagePanel />;
