+++ src/components/passwords/password-import-parsers.test.ts
@@ -7,7 +7,7 @@ import {
   parseCsv,
   parseJson,
   formatLabels,
-} from "./import-dialog-parsers";
+} from "./password-import-parsers";
 
 // ─── parsePasswdSsoPayload ───────────────────────────────────
