+++ src/components/passwords/password-import-tags.test.ts
@@ -1,6 +1,6 @@
 import { describe, expect, it, vi } from "vitest";
 import { API_PATH } from "@/lib/constants";
-import { resolveEntryTagIds, resolveTagNameToIdForImport } from "@/components/passwords/import-dialog-utils";
+import { resolveEntryTagIds, resolveTagNameToIdForImport } from "@/components/passwords/password-import-utils";
 
 type ImportEntry = Parameters<
   typeof resolveTagNameToIdForImport
